Error page refactor
