+++ public/assets/locales/en/errors.json
@@ -1,5 +1,6 @@
 {
   "It looks like you found a glitch in the matrix...": "It looks like you found a glitch in the matrix...",
-  "Page Not Found": "Page Not Found",
-  "404": "404"
+  "Aquarium not found": "Aquarium not found",
+  "Unknown error": "Unknown error",
+  "404": "Page not found"
 }
