+++ public/assets/locales/pl/configuration.json
@@ -1,8 +1,4 @@
 {
-  "changes prompt": "czy zatwierdzić zmiany?",
-  "apply": "zatwierdź",
-  "cancel": "anuluj",
-  "Back to Dashboard": "Powrót do strony startowej"
   "changes prompt": "Czy zatwierdzić zmiany?",
   "apply": "Zatwierdź",
   "cancel": "Anuluj",
