+++ public/assets/locales/pl/errors.json
@@ -1,5 +1,6 @@
 {
   "It looks like you found a glitch in the matrix...": "Wygląda na to, że znalazłeś błąd w matrixie...",
-  "Page Not Found": "Nie znaleziono strony",
-  "404": "404"
+  "Aquarium not found": "Nie znaleziono akwarium",
+  "Unknown error": "Nieznany błąd",
+  "404": "Nie znaleziono strony"
 }
