+++ src/pages/error/Error.tsx
@@ -4,21 +4,28 @@ import { useTranslation } from 'react-i18next';
 import { faArrowLeft } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
-import './404.scss';
+import './Error.scss';
 
-const PageNotFound = () => {
-  const { t } = useTranslation();
+type Props = {
+  statusCode: number;
+  message?: string;
+};
+
+const Error = ({ statusCode, message }: Props) => {
+  const { t } = useTranslation(['errors']);
+
+  const translatedMessage = t(message || statusCode.toString());
 
   return (
     <Col className="text-center">
-      <div className="error mx-auto shake">{t('404', { ns: 'errors' })}</div>
+      <div className="error mx-auto shake">{statusCode}</div>
       <p className="lead text-gray-800 mb-5">
-        {t('Page Not Found', { ns: 'errors' })}
+        {translatedMessage === statusCode.toString()
+          ? t('Unknown error')
+          : translatedMessage}
       </p>
       <p className="lead text-gray-500 mb-0">
-        {t('It looks like you found a glitch in the matrix...', {
-          ns: 'errors',
-        })}
+        {t('It looks like you found a glitch in the matrix...')}
       </p>
       <Nav.Item className="link text-primary">
         <Nav.Link className="text-primary" href="/dashboard">
@@ -30,4 +37,4 @@ const PageNotFound = () => {
   );
 };
 
-export default PageNotFound;
+export default Error;
