+++ src/pages/index.ts
@@ -1,4 +1,4 @@
-export { default as PageNotFound } from './404/404';
+export { default as Error } from './error/Error';
 export { default as Working } from './working/Working';
 export { default as Settings } from './settings/Settings';
 export { default as Edit } from './edit/Edit';
